third_party/blink/web_tests/http/tests/worklet/webexposed/global-interface-listing-layout-worklet.html

<!DOCTYPE html>
<html>
<body>
<script>
  if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
  }

  console.log("This test logs exposed APIs from LayoutWorkletGlobalScope");

  if ("layoutWorklet" in CSS) {
    CSS.layoutWorklet.addModule("/js-test-resources/global-interface-listing.js")
      .then(function() {
        if (window.testRunner) {
          testRunner.notifyDone();
        }
    });
  } else {
    console.log("CSS.layoutWorklet not exposed");
    if (window.testRunner) {
      testRunner.notifyDone();
    }
  }
</script>
</body>
</html>
