third_party/blink/web_tests/virtual/stable/http/tests/worklet/webexposed/global-interface-listing-layout-worklet-expected.txt

CONSOLE MESSAGE: This test logs exposed APIs from LayoutWorkletGlobalScope
CONSOLE MESSAGE: CSS.layoutWorklet not exposed
